Created the 'ps2controller_2_0' sketch.
This implements the new controller layout we discussed using a
centralized 'motors' function to translate throttle, yaw, collective
(vertical throttle) and pitch values from -100 to 100 to the appropriate
motor outputs.

Yaw and Pitch are achieved by reducing power on one thruster. This
sketch does not let thrusters run in opposite directions in an attempt
to reduce some of the adverse handling characteristics seen the pool. We
may have to revisit this.